refactor(orchestrator): apply dedup budget in place in cache.go to minimize diff
